moved share button
